[vim] Fix glitchy autocomplete/snippet integration behavior
The author of `ncm2/ncm2-neosnippet` fixed the bug that was preventing
snippets from expanding after complete.  Now I've run into another bug
(https://github.com/Shougo/neosnippet.vim/issues/455) but that's a minor
annoyance.

This change abandons the idea of using TAB to navigate snippets (I've
already gotten used to C-k), but expands the possible scenarios to cover
completions from ncm2, completions from neosnippet, and non-snippets for
which C-k was accidentally pressed in the heat of the moment.  This is
_almost_ the ideal config.