[polybar] fix tray size and inconsistent position
Polybar can only display the tray on one monitor; which one gets it is
basically a race condition.  I've changed the launch script now to
always place the tray on the first monitor in the `xrandr` output.

I've also set the tray size to `32` so the icons are much more legibile
on a HiDPI screen