Disable `Nerd Font` variant of `Noto Serif`
I was using the `NotoSerif NerdFont` variation of `Noto Serif` because I
already use the nerdfont variant for the terminal and I figured I may as
well be consistent and make sure I can support NerdFont glyphs in the
browser also.  It was a good idea, and someday maybe it will work.

But for now there's this issue [1], which is blocking any progress.
It's maddening to read text when various emojis are injected in place of
certain ligatures.

[1](https://github.com/ryanoasis/nerd-fonts/issues/254)