[tmux] Fix bug in session switcher script
On at least one occassion I've encountered a bug in tmux in which the
same session is listed twice in `tmux list-windows`.  It becomes
impossible to switch to that session using the `ftwind` script as a
result.

The fix is to pass the list of matching windows through `head -n 1` to
get the first one instead of matching all of them.