[vim] Experment with Rust Analyzer instead of RLS with Vim
I ended up not proceeding with Rust Analyzer, which is why this logic is
commented out in the commit.  But that's because I don't think
`LanguageClient-neovim` is the right plugin.  From the rust-analyzer
docs it's much more full featured than RLS and is where most of the
development effort is going, so I should be using it.  But it seems the
Rust core team members who use Vim use `coc.vim` instead, so I need to
switch to that first.