[tmux] Disable underline color and underline style for now
This is, as far as I can tell, not at all supported in alacritty.

You can test in the future with the following shell command:

```
$ printf "\x1b[4:3m\x1b[58:2:251:73:52m underlined\n"
```

On Kitty that prints out the word "underlined", with a red squiggly
underline.  On Alacritty it prints out the raw escape codes.

Strangely, even with tmux configured to emit the underscore color
sequence, neovim still doesn't show underlines properly; it renders the
raw escape sequence.  I think there's something wrong with the tmux
config related tu `Setulc`.