[i3] update i3 config for fedora 32
This means no i3-gaps, chromium instead of chrome, plus a few changes I
made in the sway config that I want to port to i3.

Sadly, wayland in general and sway in particular are not yet ready for
use as a daily driver.  The main issue is lack of Zoom screen sharing
support.