[neovim] Start a new config just for neovim
Now that NeoVim 0.5 is out I wanted to try using the native LSP support.
Unfortunately it's still lacking compared to CoC with regards to Rust
support, so I'm going to stick with what I CoC for now, but I did make
progress in setting up the more NeoVim-native config so I'm leaving it
here as `init-neovim.vim` for when the tooling matures a bit.