[vim] Fzf in vim now uses tmux popups
This required an annoying hack for `rg` because fzf-vim assumes that
`rg` is in the path, but in tmux popups the normal shell init scripts
aren't evaluated so things like `rg` that are installed in
`~/.cargo/bin` were not in the path, alghouth ~/.local/bin is.  So I
made a shitty wrapper script in ~/.local/bin/rg which invokes
~/.cargo/bin/rg.  That's really annoying, hopefully there's a better
solution forthcoming.