[nvim] Update coc settings and bindings for latest version
The latest update changed how the auto-complete popup worked, and it
broke my workflow.  It would be nice to be able to use the new approach
with automatically selects the first matching suggestion, but the
binding of <CR> to actually perform that completion for some reason
doesn't work.  It's probably an issue in my config but I can't find it
and I'm already used to the old way where I have to explicitly scroll
down one to match the suggested completion, so I can live with it the
way it is.