[tmux] update `ftwind` script to be more user friendly
Now the tmux popup is bigger, and the tmux session ID is moved to the
first column.  This is very helpful when there are a lot of sessions
with similar names, so I can type `'@` followed by the session ID and
immediately jump to the session I want with no scrolling.

I also fixed an annoying bug whereby if I hit ESC and dismiss the tmux
popup, tmux would switch to the login shell.  Now dismissing the tmux
popup will not change the current window at all.