[vim] recreate neovim config with updated plugins and Lua config language
This has been a long time coming.  Much of my vim comfig is two or more
years old, and a lot has happened in that time.  Some plugins that I
used everyday are no longer maintained or have been replaced with better
alternatives using the more modern and perfromant Lua language and
built-in neovim features.  This is a work in progress commit, I'm not
ready to switch to this new config full time yet.