[neovim] Experiment w/ some changes to Copilot config and add SourceGraph
This is experimental.  Early results suggest that the integration of the
Copilot.lua plugin into Neovim's completions is flaky.  When I cycle
between Copilot suggestions, it doesn't reliably delete all of the first
suggestion, nor does it intent it properly, so the result is shitty.

I'm committing this for the record but I plan to tweak it further.

I also added but did not yet do anything with the Sourcegraph plugin, as
I'm getting some value out of Cody in VS Code and don't want to get
stuck there just because the tooling is better.