[neovim] Remove sourcegraph plugin which is completely broken
It seems like this plugin assumes that Neovim is run on a local system
w/ a window manager and a browser, so it can launch some auth flow.  I
tried running `SourcegraphLogin!` which lets me paste in an auth token
manually, but it silently does nothing, and after restart is still not
authenticated.

Sourcegraph warns that this is "experimental"; I would go further and
say that it's not something they are investing in at all so if you want
a tool that works in Neovim you should stick with Copilot and hope it
improves.