[vim] Use an OSC52 plugin so I can yank text into the client terminal clipboard
This is a huge help.  I normally run Neovim on a remote server in a data
center with reliable power and fast internet.  Sometimes I want to copy
some text from within neovim and paste it into a browser window or
terminal on my local system.  I didn't have an easy way to do this
before.

Note that I had to change some tmux config to support this as well.
That will be in a separate commit.