[tmux] Enable passthrough of escape sequences to the terminal
This is needed to allow the OSC52 plugin in neovim to write OSC 52
escape sequences when running in tmux, and not have tmux intercept them.

This works by having a special escape sequence that activates this
passthrough mode.  The OSC52 plugin knows about this, and is configured
to use it.

That means the OSC52 functionality in Neovim won't work if it's not
running under tmux, but I always run terminal programs under tmux, and
if I don't it must be running locally in which case I can just yank to
the system clipboard and don't need the OSC52 functionality at all.