[nvim] Reset neovim config with mostly-stock LazyVim
In this branch I'm experimenting with LazyVim as the basis for my
config, to modernize things without a lot of effort and tinkering.  So
far it's more frustrating than useful but I'm going to keep poking at
it.

Subsequent commits will consist of me applying my previous config on top
of LazyVim.