[nvim] Port over more of my configs from the pre-LazyVim version
Here I brought in the cmp config (or some of it anyway), and adapted my
older Rust config to use whatever the new neovim hottness is.  Sadly, it
seems like keeping one's NeoVim config up to date w/ the latest hotness
that the cool kids are using is itself a full time job.  For now I'll
put this down to the fact that LSP support in Neovim is itself pretty
new and still a moving target.  If the state of the art for Rust support
in Neovim completely changes for a fourth time then I'll have to revisit
my life choices.

I also adapted some treesitter and mason configs.