[nvim] Fix completion and copilot config
The move to LazyVim required a lot more fiddling than I expected, in
order to get auto-completion and Copilot working the way I like.  But
I think I've finally got it.

I'm shocked that there are people who use Copilot as a completion
source, as opposed to a dedicated suggestion source.  Madness.