[ssh] Enable SSH command-line by default and add mac identity agent
Starting in OpenSSH 9.2 (https://www.openssh.com/releasenotes.html),
the SSH command line is off by default.  That sucks because I use it to
add new tunnels to an existing SSH connection.

So this enables that by default.

This also adds a change that I've had on my mac for a long time but
didn't commit because I couldn't figure out how to make it conditional.
It will use the 1Password SSH agent on my macbook, which is a nice
improvement over using the macOS keychain which is what I did before.