[tmux] Add Wezterm and tidy up some things
I'm switching to Wezterm, which supports undercurls and colored
underlines.  It turns out that by new Alacritty does also, although I
didn't realize that.  So this sets the appropriate terminal overrides to
enable this.

Also activate OSC 7 support, so the shells can report current directory
to the terminal, and enable set-titles so that tmux passes a meaningful
title to the terminal emulator.

Overline support still isn't working right but I don't think it ever was
so I don't want to waste time on it now.