[ranger] Clean up ranger config
There were some warnings loading the config; I suppose it was old.  I
thought this was causing a problem w/ ranger running on mac, but it's
not.  It doesn't like to see `tmux-256color` term, even though that
works fine on Linux and there is a terminfo entry for that terminal.