[nvim] After update to lazyvim 0.14, force return to nvim-cmp
I hate that this was done unilaterally, and I hate that going back to
nvim-cmp requires more fucking around than just enabling the extra in
the config.  I think long-term LazyVim is not for me, it's for people
who will take whatever LazyVim gives them and not try to deviate too
much.  I'm not that person, but I am the person who doesn't have time to
rip out lazyvim and reconstruct my current config independent of it.