[nvim] Update copilot.lua config to use latest example config
There was a refactor with breaking changes a few days ago.  I'm not sure
what those were because the details were not documented, I just saw it
in the commit log.  I updated this based on the current docs on the site
for copilot.lua, hopefully it doesn't break or regress anything.