[nvim] Fix broken default Python beautifier settings
It was using `pylsp`'s internal formatter which is broken in ways that
would convert valid, running Python code into invalid code that fails to
parse.  Why that was ever considered a reasonable default is beyond me.

This disables all pylsp formatting features, and sets `black` as the
Python formatter.  This assumes that `black` is installed, which
hopefully it will be.