[zsh] Ensure ~/.local/bin is always at the front of PATH
This should have been done originally but I didn't notice until
recently.

I want to make a script `gemini` that wraps the NPM-installed CLI of the
same name.  Only to find out that the NPM packages dir was earlier in
the path than ~/.local/bin.

This (admittedly kludgy) change keeps the .profile entry that sets
~/.local/bin at the front of the PATH to ensure that it's always done in
all shells interactive or otherwise, but then at the very end of .zshrc
edits PATH to ensure it's always at the front.

Stupid, I know.